requirements.txt
fastapi
uvicorn
requests
jinja2
firebase-admin
twilio
python-dotenv
